requirement.txt
# Function dependencies, for example:
# package>=version
google-cloud-bigquery
googleapis-common-protos
Flask==1.0.2